Fix when not creating index with CRAM (#3324)

+++ modules/nf-core/gatk4/markduplicates/main.nf
@@ -45,7 +45,7 @@ process GATK4_MARKDUPLICATES {
         $args
 
 
-    if  [[ ${prefix} == *.cram ]]; then
+    if  [[ ${prefix} == *.cram ]]&&[[ -f ${prefix}.bai ]]; then
         mv ${prefix}.bai ${prefix}.crai
     fi
